detector/requirements.txt
opencv-python-headless>=4.9.0
numpy>=1.26.0
redis>=5.0.0
psycopg2-binary>=2.9.9
minio>=7.2.0
Pillow>=10.2.0
schedule>=1.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
